Fix: Zipcode and house number not saved (github #74)
Zipcode is either stored as "zip_code" and sometimes "postcode"
plus street number is "street_nr" and "street_no".
This leads to confusion even for the plugin
especially when it comes down to input fields
and storing data.
We need to bring in some compatibility layer
until version 2.0.0 where those shall be removed.