Sanitize from address before submitting to API (github #41)
The user can define a default sender address
in the shipcloud.io backend
and does not need to fill out the shipcloud config in WooCommerce.
Unfortunately the WooCommerce settings have a default for the country
so the from-data always contain a country
making the API consider the from-address as incomplete.
To suppress this the from address will be sanitized before submitting
removing the from address when it only contains the country.