Hotfix: Use secondary address as care-of (github #99)
Customers can enter a secondary address in the billing address
while the admin is capable of providing some "care_of" address
containing the apartment or some other post-box to deliver to.
Those two are almost the same but didn't help each another
or were connected.
From now on an secondary billing address will be used for the care-of
address.

- Another hotfix is in this issue regarding correct escaping.