Carrier select: Fixed preselect of carrier data (github #66)
The parcel template creation con configure which carrier,
service and package type will be used.
This wasn't filled completely when entering the form,
due to the fresh introduced package type
and JavaScript form.
So the JavaScript form got a new method for selecting items
and this shall be injected with data when entering/building the form.